// src/app/components.enum.ts
export enum AppComponents {
    gas = 'gas', // Газопроникність
    press = 'press', // Капілярний тиск
    centrifuge = 'centrifuge', // Центрифугування
    resistant = 'resistant', // Питомий опір
    pores = 'pores', // Пористість
    volume = 'volume', // Об'ємна вага
    density = 'density', // Мінералогічна густина
    carbon = 'carbon', // Карбонатність
    oil = 'oil', // Нафтонасичення
    organic = 'organic', // С Органічний
    wax = 'wax', // 'Об.вага в парафіні'
}